Remove prepend of account name to custom service URL
If one wants to use https://blob.core.windows.net/, or some other non-default URL to Azure Storage,
the host part of the URL will be prepended with the storage account of the build.

Example:

1. The blob service endpoint URL is set to https://blob.core.windows.net for account 'foo-account'
2. First build publishes its artefacts successfully to https://foo-account.blob.core.windows.net
3. Second build fails to publish the build artifacts with the following error message:

    ```
    Archiving artifacts
    Failed to validate storage account details. Please verify storage account name and key. If you are using a private or other Windows Azure cloud service, make sure that the blob endpoint url is correct.
    Storage Account name --->foo-account<----
    Blob end point url --->https://foo-account.foo-account.blob.core.windows.net/<----
    Build step 'Upload artifacts to Microsoft Azure Blob storage' changed build result to UNSTABLE
    Finished: UNSTABLE
    ```

This commit removes the prepend of the storage account name in the Utils.getBlobEP method and updates its usages.
Lacking an available development environment, tests for the new behavior were not added (help wanted with this).