Upload and download files with relative paths
When we originally were uploading to azure we were flattening the directories.  This meant that two files in different directories with the same name would overwrite each other.  The regular artifact copying plugin avoids this and uploads to the server using relative paths.  Add an option to flatten directories on download.