Added support for HTTP proxy settings coming from OS environment variables http_proxy, https_proxy, and no_proxy (the standard linux format) in case where the java properties would not have been set otherwise. Indeed Jenkins does not support HTTP proxy setup apart from plugin update :(
Note: these settings are shared with all the JVM so be sure that they are configured correctly otherwise other plugins might stop reaching their HTTP destinations ! For example if another plugin was poking an address on the intranet, and this address is not properly declared in no_proxy, then this address will now be resolved using the proxy and will fail. Classic to people using proxies, but good to be aware of.